Canonical arithmetic expressions and recursion-closure querying
Replace the add/subtract/multiply/divide Expression enum with a canonical
normal form (sum of products of powers): commutative operands sorted,
constants folded, like terms collected, so logically-equal expressions are
also structurally equal and dedup on the _rule UNIQUE constraint. Subtract,
divide, and negation lower to add/multiply/exponent.

Add query-time recursive-closure building: self-recursive predicates become
WITH RECURSIVE CTEs while dependent predicates are inlined as streaming
FROM-subqueries, with the query's equality constraints propagated down the
dependency chain to bound (and terminate) each recursive step. Body-side
arithmetic is bound by inverting the expression against its column.

Switch Term numbers from Float to Double for full integer precision and
plain-decimal rendering (scientific notation broke recursion-bound
extraction). Emit SELECT DISTINCT so projected queries return sets. Extract
the SQLite scalar functions into SQLiteFunctions.swift.

+++ Tests/RBDBTests/TermTests.swift
@@ -22,7 +22,7 @@ import Testing
 	try assertJSON(Term.number(5.7), expect: "{\"\":5.7}")
 	try assertJSON(Term.number(-5.123), expect: "{\"\":-5.123}")
 	#expect(throws: EncodingError.self) {
-		try assertJSON(Term.number(Float.nan), expect: "")
+		try assertJSON(Term.number(Double.nan), expect: "")
 	}
 }
